+++ src/main/java/com/salesforce/storm/spout/dynamic/DelegateSpout.java
@@ -29,7 +29,7 @@ import com.salesforce.storm.spout.dynamic.consumer.ConsumerState;
 import com.salesforce.storm.spout.dynamic.consumer.Consumer;
 
 /**
- * A virtualized spout that is run from within the DyanmicSpout by the SpoutCoordinator.
+ * A virtualized spout that is run from within the DyanmicSpout by the Coordinator.
  */
 public interface DelegateSpout {
