+++ src/main/java/com/salesforce/storm/spout/dynamic/coordinator/SpoutMonitor.java
@@ -26,13 +26,12 @@
 package com.salesforce.storm.spout.dynamic.coordinator;
 
 import com.salesforce.storm.spout.dynamic.Tools;
-import com.salesforce.storm.spout.dynamic.MessageId;
+import com.salesforce.storm.spout.dynamic.VirtualSpoutCoordinator;
 import com.salesforce.storm.spout.dynamic.VirtualSpoutIdentifier;
 import com.salesforce.storm.spout.dynamic.config.SpoutConfig;
 import com.salesforce.storm.spout.dynamic.DelegateSpout;
 import com.salesforce.storm.spout.dynamic.VirtualSpout;
 import com.salesforce.storm.spout.dynamic.metrics.MetricsRecorder;
-import com.salesforce.storm.spout.dynamic.buffer.MessageBuffer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -72,32 +71,12 @@ public class SpoutMonitor implements Runnable {
     private final Queue<DelegateSpout> newSpoutQueue;
 
     /**
-     * This buffer/queue holds tuples that are ready to be sent out to the topology.
-     * It is filled by VirtualSpout instances, and drained by DynamicSpout.
+     * Routes messages from VirtualSpouts to the Coordinator.
      */
-    private final MessageBuffer tupleOutputQueue;
+    private final VirtualSpoutCoordinator virtualSpoutCoordinator;
 
     /**
-     * This buffer/queue holds tuples that are ready to be acked by VirtualSpouts.
-     * Its segmented by VirtualSpout ids => Queue of tuples to be acked.
-     * It is filled by DynamicSpout, and drained by VirtualSpout instances.
-     */
-    private final Map<VirtualSpoutIdentifier,Queue<MessageId>> ackedTuplesQueue;
-
-    /**
-     * This buffer/queue holds tuples that are ready to be failed by VirtualSpouts.
-     * Its segmented by VirtualSpout ids => Queue of tuples to be failed.
-     * It is filled by DynamicSpout, and drained by VirtualSpout instances.
-     */
-    private final Map<VirtualSpoutIdentifier,Queue<MessageId>> failedTuplesQueue;
-
-    /**
-     * This buffer/queue holds errors that should be reported up to the topology.
-     */
-    private final Queue<Throwable> reportedErrorsQueue;
-
-    /**
-     * This latch allows the SpoutCoordinator to block on start up until its initial
+     * This latch allows the Coordinator to block on start up until its initial
      * set of VirtualSpout instances have started.
      */
     private final CountDownLatch latch;
@@ -142,10 +121,6 @@ public class SpoutMonitor implements Runnable {
     /**
      * Constructor.
      * @param newSpoutQueue Queue monitored for new Spouts that should be started.
-     * @param tupleOutputQueue Queue for pushing out Tuples to the topology.
-     * @param ackedTuplesQueue Queue for incoming Tuples that need to be acked.
-     * @param failedTuplesQueue Queue for incoming Tuples that need to be failed.
-     * @param reportedErrorsQueue Queue for any errors that should be reported to the topology.
      * @param latch Latch to allow startup synchronization.
      * @param clock Which clock instance to use, allows injecting a mock clock.
      * @param topologyConfig Storm topology config.
@@ -153,20 +128,14 @@ public class SpoutMonitor implements Runnable {
      */
     public SpoutMonitor(
         final Queue<DelegateSpout> newSpoutQueue,
-        final MessageBuffer tupleOutputQueue,
-        final Map<VirtualSpoutIdentifier, Queue<MessageId>> ackedTuplesQueue,
-        final Map<VirtualSpoutIdentifier, Queue<MessageId>> failedTuplesQueue,
-        final Queue<Throwable> reportedErrorsQueue,
+        final VirtualSpoutCoordinator virtualSpoutCoordinator,
         final CountDownLatch latch,
         final Clock clock,
         final Map<String, Object> topologyConfig,
         final MetricsRecorder metricsRecorder
     ) {
         this.newSpoutQueue = newSpoutQueue;
-        this.tupleOutputQueue = tupleOutputQueue;
-        this.ackedTuplesQueue = ackedTuplesQueue;
-        this.failedTuplesQueue = failedTuplesQueue;
-        this.reportedErrorsQueue = reportedErrorsQueue;
+        this.virtualSpoutCoordinator = virtualSpoutCoordinator;
         this.latch = latch;
         this.clock = clock;
         this.topologyConfig = Tools.immutableCopy(topologyConfig);
@@ -274,9 +243,7 @@ public class SpoutMonitor implements Runnable {
 
                 final SpoutRunner spoutRunner = new SpoutRunner(
                     spout,
-                    tupleOutputQueue,
-                    ackedTuplesQueue,
-                    failedTuplesQueue,
+                    virtualSpoutCoordinator,
                     latch,
                     getClock(),
                     getTopologyConfig()
@@ -359,10 +326,10 @@ public class SpoutMonitor implements Runnable {
             getNumberOfFailedTasks(),
             executor.getTaskCount()
         );
-        logger.info("MessageBuffer size: {}, Running VirtualSpoutIds: {}", tupleOutputQueue.size(), spoutRunners.keySet());
+        logger.info("MessageBuffer size: {}, Running VirtualSpoutIds: {}", virtualSpoutCoordinator.messageSize(), spoutRunners.keySet());
 
         // Report to metrics record
-        getMetricsRecorder().assignValue(getClass(), "bufferSize", tupleOutputQueue.size());
+        getMetricsRecorder().assignValue(getClass(), "bufferSize", virtualSpoutCoordinator.messageSize());
         getMetricsRecorder().assignValue(getClass(), "running", executor.getActiveCount());
         getMetricsRecorder().assignValue(getClass(), "queued", executor.getQueue().size());
         getMetricsRecorder().assignValue(getClass(), "errored", getNumberOfFailedTasks());
@@ -503,19 +470,12 @@ public class SpoutMonitor implements Runnable {
     }
 
     /**
-     * @return The Error Queue.
-     */
-    Queue<Throwable> getReportedErrorsQueue() {
-        return reportedErrorsQueue;
-    }
-
-    /**
      * Adds an error to the reported errors queue.  These will get pushed up and reported
      * to the topology.
      * @param throwable The error to be reported.
      */
     private void reportError(final Throwable throwable) {
-        getReportedErrorsQueue().add(throwable);
+        virtualSpoutCoordinator.publishError(throwable);
     }
 
     /**
